+++ Runtime/BaseComponent.cs
@@ -2,6 +2,14 @@
 
 namespace Lazlo.Gocs
 {
+	/// <summary>
+	/// The base class from which to derive components.
+	/// </summary>
+	/// <remarks>
+	/// You can also manually implement <see cref="IComponent"/> if you
+	/// cannot derive from this class.
+	/// </remarks>
+	/// <seealso cref="BaseImplementation"/>
 	public abstract class BaseComponent : MonoBehaviour, IComponent
 	{
 		protected virtual void Awake()