+++ Runtime/BaseImplementation.cs
@@ -3,6 +3,12 @@ using UnityObject = UnityEngine.Object;
 
 namespace Lazlo.Gocs
 {
+	/// <summary>
+	/// Provides helpers to implement lifecycle events on components and systems.
+	/// Use this class when creating custom base classes from interfaces to ensure future-proofing.
+	/// </summary>
+	/// <seealso cref="IComponent"/>
+	/// <seealso cref="ISystem"/>
 	public static class BaseImplementation
 	{
 		#region Component