+++ Runtime/Event_TArgs.cs
@@ -4,20 +4,40 @@ using UnityEngine;
 
 namespace Lazlo.Gocs
 {
+	/// <summary>
+	/// Represents an event with arguments.
+	/// Events support assigning handlers and invoking from anywhere. 
+	/// </summary>
+	/// <remarks>
+	/// Use an arguments struct or class if you need to pass multiple arguments.
+	/// </remarks>
+	/// <typeparam name="TArgs">The type of argument.</typeparam>
+	/// <seealso cref="Event"/>
 	public sealed class Event<TArgs>
 	{
 		private readonly HashSet<Action<TArgs>> handlers;
 
+		/// <summary>
+		/// Creates a new event without any handler.
+		/// </summary>
 		public Event()
 		{
 			handlers = new HashSet<Action<TArgs>>();
 		}
 
+		/// <summary>
+		/// Creates a new event with a specified handler handler.
+		/// </summary>
+		/// <param name="handler">The event handler.</param>
 		public Event(Action<TArgs> handler) : this()
 		{
 			AddHandler(handler);
 		}
 
+		/// <summary>
+		/// Adds a handler to the event.
+		/// </summary>
+		/// <param name="handler">The event handler.</param>
 		public void AddHandler(Action<TArgs> handler)
 		{
 			if (handler == null)
@@ -28,6 +48,10 @@ namespace Lazlo.Gocs
 			handlers.Add(handler);
 		}
 
+		/// <summary>
+		/// Removes a handler from the event.
+		/// </summary>
+		/// <param name="handler">The event handler.</param>
 		public void RemoveHandler(Action<TArgs> handler)
 		{
 			if (handler == null)
@@ -38,6 +62,10 @@ namespace Lazlo.Gocs
 			handlers.Remove(handler);
 		}
 		
+		/// <summary>
+		/// Invokes every handler on the event.
+		/// </summary>
+		/// <param name="args">The event arguments.</param>
 		public void Invoke(TArgs args)
 		{
 			foreach (var handler in handlers)