+++ Runtime/ISystem.cs
@@ -1,5 +1,11 @@
 ﻿namespace Lazlo.Gocs
 {
+	/// <summary>
+	/// Represents a system.
+	/// Systems are responsible for handling common logic and event dispatch on batches of components.
+	/// To receive callbacks on component creation and destruction, see <see cref="IWorldCallbackReceiver"/>.
+	/// </summary>
+	/// <seealso cref="BaseSystem"/>
 	public interface ISystem
 	{