+++ Runtime/SystemEvents_Args.cs
@@ -3,10 +3,25 @@ using System.Collections.Generic;
 
 namespace Lazlo.Gocs
 {
+	/// <summary>
+	/// Manages event handler registration for a system.
+	/// This version supports event arguments.
+	/// </summary>
+	/// <seealso cref="SystemEvents"/>
 	public sealed class SystemEvents<TArgs>
 	{
 		private readonly Dictionary<Event<TArgs>, Action<TArgs>> handlers = new Dictionary<Event<TArgs>, Action<TArgs>>();
-
+		
+		/// <summary>
+		/// Adds or removes a system event handler for the specified event.
+		/// In <see cref="IWorldCallbackReceiver.OnCreatedComponent"/>, pass a handler to add it to the event.
+		/// In <see cref="IWorldCallbackReceiver.OnDestroyingComponent"/>, pass null to remove the handler from the event.
+		/// </summary>
+		/// <remarks>
+		/// System events can only add one handler per event.
+		/// </remarks>
+		/// <param name="event">The event on which to add or remove the handler.</param>
+		/// <returns>The event handler.</returns>
 		public Action<TArgs> this[Event<TArgs> @event]
 		{
 			get